Move `_remote_learn` to `AbsTrainWorker`.
